# IGI_API/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the IGI API server documentation.",
        "title": "IGI_API",
        "contact": {},
        "version": "1.0.1"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/people": {
            "get": {
                "description": "Retrieve all people data from SWAPI with pagination support.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "People"
                ],
                "summary": "Get all people",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default: 15)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of People with pagination",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/IGI_API_internal_models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "response": {
                                            "$ref": "#/definitions/IGI_API_internal_models.SWAPIList"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/IGI_API_internal_models.APIResponse"
                        }
                    }
                }
            }
        },
        "/people/{id}": {
            "get": {
                "description": "Retrieve a specific person's data by their ID from SWAPI.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "People"
                ],
                "summary": "Get person by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Person ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Person Data",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/IGI_API_internal_models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "response": {
                                            "$ref": "#/definitions/IGI_API_internal_models.PeopleByIDResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid Person ID",
                        "schema": {
                            "$ref": "#/definitions/IGI_API_internal_models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/IGI_API_internal_models.APIResponse"
                        }
                    }
                }
            }
        },
        "/planet": {
            "get": {
                "description": "Retrieve paginated planet data from SWAPI.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Planets"
                ],
                "summary": "Get all planets",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number (default is 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default is 15)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of Planets with pagination details",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/IGI_API_internal_models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "response": {
                                            "$ref": "#/definitions/IGI_API_internal_models.SWAPIList"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid Query Parameters",
                        "schema": {
                            "$ref": "#/definitions/IGI_API_internal_models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/IGI_API_internal_models.APIResponse"
                        }
                    }
                }
            }
        },
        "/planet/{id}": {
            "get": {
                "description": "Retrieve a specific planet's data by its ID from SWAPI.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Planets"
                ],
                "summary": "Get planet by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Planet ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Planet Data",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/IGI_API_internal_models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "response": {
                                            "$ref": "#/definitions/IGI_API_internal_models.PlanetByIDResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid Planet ID",
                        "schema": {
                            "$ref": "#/definitions/IGI_API_internal_models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/IGI_API_internal_models.APIResponse"
                        }
                    }
                }
            }
        },
        "/search": {
            "get": {
                "description": "Search for people or planets using a keyword from SWAPI.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Search"
                ],
                "summary": "Search resources by keyword",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search Keyword",
                        "name": "keyword",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Data Source (people or planets)",
                        "name": "source",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Page number (default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Items per page (default: 15)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field (name or height | default: name)",
                        "name": "sortBy",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order (asc or desc | default: asc)",
                        "name": "sortOrder",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful search operation",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/IGI_API_internal_models.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "response": {
                                            "$ref": "#/definitions/IGI_API_internal_models.SearchResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid Query Parameters",
                        "schema": {
                            "$ref": "#/definitions/IGI_API_internal_models.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/IGI_API_internal_models.APIResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "IGI_API_internal_models.APIResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "response": {
                    "description": "this can be any type of data. Data is presented as a property of response"
                },
                "status": {
                    "type": "integer"
                }
            }
        },
        "IGI_API_internal_models.PeopleByIDResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {
                    "type": "object",
                    "properties": {
                        "description": {
                            "type": "string"
                        },
                        "properties": {
                            "$ref": "#/definitions/IGI_API_internal_models.PersonProperties"
                        },
                        "uid": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "IGI_API_internal_models.PersonProperties": {
            "type": "object",
            "properties": {
                "birth_year": {
                    "type": "string"
                },
                "created": {
                    "type": "string"
                },
                "edited": {
                    "type": "string"
                },
                "eye_color": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "hair_color": {
                    "type": "string"
                },
                "height": {
                    "type": "string"
                },
                "homeworld": {
                    "type": "string"
                },
                "mass": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "skin_color": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "IGI_API_internal_models.PlanetByIDResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {
                    "type": "object",
                    "properties": {
                        "description": {
                            "type": "string"
                        },
                        "properties": {
                            "$ref": "#/definitions/IGI_API_internal_models.PlanetProperties"
                        },
                        "uid": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "IGI_API_internal_models.PlanetProperties": {
            "type": "object",
            "properties": {
                "climate": {
                    "type": "string"
                },
                "created": {
                    "type": "string"
                },
                "diameter": {
                    "type": "string"
                },
                "edited": {
                    "type": "string"
                },
                "gravity": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "orbital_period": {
                    "type": "string"
                },
                "population": {
                    "type": "string"
                },
                "rotation_period": {
                    "type": "string"
                },
                "surface_water": {
                    "type": "string"
                },
                "terrain": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "IGI_API_internal_models.SWAPIList": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "next": {
                    "type": "string"
                },
                "previous": {},
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/IGI_API_internal_models.SWAPIListItem"
                    }
                },
                "total_pages": {
                    "type": "integer"
                },
                "total_records": {
                    "type": "integer"
                }
            }
        },
        "IGI_API_internal_models.SWAPIListItem": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "IGI_API_internal_models.SWAPISearchItemPeople": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "properties": {
                    "$ref": "#/definitions/IGI_API_internal_models.PersonProperties"
                },
                "uid": {
                    "type": "string"
                }
            }
        },
        "IGI_API_internal_models.SWAPISearchItemPlanet": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "properties": {
                    "$ref": "#/definitions/IGI_API_internal_models.PlanetProperties"
                },
                "uid": {
                    "type": "string"
                }
            }
        },
        "IGI_API_internal_models.SearchResponse": {
            "type": "object",
            "properties": {
                "dataSource": {
                    "type": "string"
                },
                "hitCount": {
                    "type": "object",
                    "properties": {
                        "total": {
                            "type": "integer"
                        },
                        "totalPagePeople": {
                            "type": "integer"
                        },
                        "totalPagePlanet": {
                            "type": "integer"
                        }
                    }
                },
                "results": {
                    "$ref": "#/definitions/IGI_API_internal_models.SearchResults"
                }
            }
        },
        "IGI_API_internal_models.SearchResults": {
            "type": "object",
            "properties": {
                "people": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/IGI_API_internal_models.SWAPISearchItemPeople"
                    }
                },
                "planets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/IGI_API_internal_models.SWAPISearchItemPlanet"
                    }
                }
            }
        }
    }
}